Implement Ctrl-o to execute a line from history and jump to the next entry
This commit replicates bashs binding to Ctrl-o.

Ctrl-o exists in bash to execute a command from the history (as Enter
does), but then display the next entry from the history. Another press
of Ctrl-o then executes that next line and displays the entry after it.

For example if the are the following entries in the history:

  echo 1
  echo 2
  echo 3

and the entry "echo 1" is the currently selected entry in re-search,
Ctrl-o would execute "echo 1" and then display "echo 2". Another press
of Ctrl-o would display "echo 3" and so on.

This allows repeating the same sequence of commands in history just by
pressing Ctrl-o multiple times.

To allow for displaying the result of the executed commands while still
inside re-search, the corresponding fish function needed to be changed
to specify the temporary file containing the line to execute as an
argument to re-search instead of always printing to stdout and using
redirection.