Better error message when history file could not be read
The error message now includes the file path and the actual error
message.

It is still not nice as the current (and more correct location
~/.local/share/fish/fish_history is tried first then the old (and wrong)
location ~/.config/fish/fish_history. If the first is not found at
all, only the second one will be mentioned in the error message.

Another still existing problem is that the paths are hardcoded and don't
use $XDG_DATA_HOME.

And the possible value of $fish_history is ignored which would overidde
the other location.