Do not reduce history on Ctrl-q
This commit changes the behaviour of Ctrl-q in that it does not remove
non-matching history entries, but instead just stores the subsearch
terms on compares them additionally to the current search term on
searching.

This is necessary to support resetting of subsearches (drop the last
subsearch) and also play nicely with the new support for Ctrl-o (as a
press of Ctrl-o should jump the next entry in the whole history, not the
next entry matching the subsearches).

Theoretically that reduces the performance of searching a bit for each
additional subsearch (as more comparisons must be done for each history
entry) whereas the previous implementation should have gotten faster for
each additional subsearch (as the number of history entries was reduced
with each subearch). However, no measurement has been done to verify
that this is actually a noticable delay.