Allow reading fish history from temporary file
The file _must_ be generated via `history --null --reverse` (probably
also restricted via `-n` to MAX_HISTORY_SIZE).
The path to the file must then be stored in the `$fish_history_file`
environment variable.

At the same time the code is refactored to multiple methods to parse
history
  - of bash
  - of fish the legacy way
  - of fish the new way

And the maximum supported commandline length is raised to 4096
characters.