Support Ctrl-a for finishing search
This basically does the same as Ctrl-e, but after that puts the cursor
at the beginning of the line (which is the normal readline behavior of
Ctrl-a).

This is done by writing the corresponding readline function into a
temporary file. Therefore it is possible to apply other readline
movement functions as well. Since a lot of those movement functions only
make sense if the cursor is at a specific position _inside_ the
commandline (e.g. M-b, M-f, C-b, C-f, etc.) a temporary file is prepared
for preparing the cursor position to put the cursor on before applying
the corresponding readline function.