[WIP] Prepare configurable prompt.
The idea is to have an environment variable (like PS1) that defines the
colors and character to use for creating the prompt. See https://github.com/jbonjean/re-search/issues/15 for a more detailed description.

This would require parsing this string each time a prompt needs to be
printed. Each character is fed to fprintf as is until a placeholder
(surrounded by ${ and }) is found. That one would be replaced by the
corresponding variable.

The problem is that some placeholders need parameter. For example the
history_item (the result of the search) has a second color to colorize
the search string.

Even more complicated would be the action as there are at the moment
four actions, each of which can have a custom text and color. And then
there should be a different color if the search term is not found.

Also we don't know whether the constant reparsing of the environment
variable and the reparsing of the placeholders + parameters have a
noticable influence on the performance.