Faster startup by avoiding writing history each time
re-search can get really nasty with long history files as the whole
history is written to disk before starting re-search. This can lead to a
very annoying lag on starting re-search (and any keypresses in that time
are being ignored).

This is now much faster since each executed command is now directly
written to a file that can be immediately read by re-search. re-search
is now as fast at startup as it was before.