Support bash and fish at the same time
No need for a compiler switch (and two binaries for the two shells)
anymore. The code now supports bash and fish in the same binary. The
shell to use must be given as environment variable RE_SEARCH_SHELL. If
set to BASH, re-search will operate on the bash history, otherwise it
will operate on the fish history.