Allow negative search terms
A new keybinding `ctrl-.` is provided to toggle „negative search“ for
finding only commandlines that /don’t/ contain the search string.

Whenn adding the search term as a subsearch term it will retain that
negative flag. It is therefore possible to combine positive and negative
search terms.

A red exclamation mark is used to indicate a negative search.